requirements.txt
# Database
psycopg2-binary==2.8

# Django
Django==3.0
djangorestframework~=3.11.0

# Authentication
djangorestframework_simplejwt==4.6.0

# Requests
requests==2.25.0

# Testing #
pytest==6.1.1
pytest-django==4.1.0
pytest-mock==3.3.1
factory-boy~=2.12.0
Faker==4.17.1

